clk: msm: clock-gcc-8996: Add 50 us delay after BLSP1 AHB clock enable
Add a new set of clk_ops for gcc_blsp1_ahb_clk with enable()
overridden to add a 50us delay after enabling the clock.
This is to workaround an issue where a clk_disable() followed
by a clk_enable() on that clock within 10 cycles can cause
instability and lead to unclocked register access on subsequent
BLSP1 register accesses.

Bug: 63095637
Change-Id: If77c366533e0d4578f1b18671604146ab985cf75
Signed-off-by: Siqi Lin <siqilin@google.com>
Signed-off-by: Ahmed Radaideh <ahmed.radaideh@gmail.com>